backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=5.1.0",
    "apscheduler>=3.11.2",
    "fastapi[standard]>=0.135.1",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.1",
    "sqlmodel>=0.0.37",
    "tavily-python>=0.7.23",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "live: requires real API keys and network access (deselect with '-m not live')",
]
